test(Toast): add demo test
