+++ example/app.json
@@ -129,9 +129,10 @@
     "van-tabbar": "./dist/tabbar/index",
     "van-tabbar-demo": "./dist/tabbar/demo/index",
     "van-tabbar-item": "./dist/tabbar-item/index",
-    "van-tag": "./dist/tag/index",packages/tag/demo/index.less,
+    "van-tag": "./dist/tag/index",
     "van-tag-demo": "./dist/tag/demo/index",
     "van-toast": "./dist/toast/index",
+    "van-toast-demo": "./dist/toast/demo/index",
     "van-transition": "./dist/transition/index",
     "van-tree-select": "./dist/tree-select/index",
     "van-datetime-picker": "./dist/datetime-picker/index",
