+++ packages/toast/demo/index.json
@@ -1,0 +1,8 @@
+{
+  "component": true,
+  "usingComponents": {
+    "van-toast": "../../toast/index",
+    "van-button": "../../button/index",
+    "demo-block": "../../../example/components/demo-block/index"
+  }
+}
